# api/docs/swagger.yaml
basePath: /
definitions:
...
  proto.CourseItemRes:
    properties:
      course_id:
        type: string
      created_at:
        type: string
      days_per_week:
        type: integer
      description:
        type: string
      duration_weeks:
        type: number
      id:
        type: string
      lesson_hours:
        type: number
      price:
        type: number
    type: object
  proto.CourseRes:
    properties:
      created_at:
        type: string
      duration:
        type: number
      id:
        type: string
      name:
        type: string
    type: object
...
  proto.CreateCourse:
    properties:
      duration:
        type: number
      name:
        type: string
    type: object
  proto.CreateCourseItem:
    properties:
      course_id:
        type: string
      days_per_week:
        type: integer
      description:
        type: string
      duration_weeks:
        type: number
      lesson_hours:
        type: number
      price:
        type: number
    type: object
...
  proto.GetListCourseItemRes:
    properties:
      course_items:
        items:
          $ref: '#/definitions/proto.CourseItemRes'
        type: array
      total_count:
        type: integer
    type: object
  proto.GetListCourseRes:
    properties:
      courses:
        items:
          $ref: '#/definitions/proto.CourseRes'
        type: array
      total_count:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for Pima
  title: Pima
  version: "1.0"
paths:
...
  /courseItems/create:
    post:
      consumes:
      - application/json
      description: Create a new course item
      parameters:
      - description: Create course item
        in: body
        name: course_item
        required: true
        schema:
          $ref: '#/definitions/proto.CreateCourseItem'
      produces:
      - application/json
      responses:
        "200":
          description: Course item created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a course item
      tags:
      - courseItems
  /courseItems/delete:
    delete:
      consumes:
      - application/json
      description: Delete a course item by ID
      parameters:
      - description: Course Item ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Course item deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a course item
      tags:
      - courseItems
  /courseItems/get:
    get:
      consumes:
      - application/json
      description: Get a course item by its ID
      parameters:
      - description: Course Item ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.CourseItemRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get course item by ID
      tags:
      - courseItems
  /courseItems/list:
    get:
      consumes:
      - application/json
      description: Get a list of course items with optional filters
      parameters:
      - description: Course ID
        in: query
        name: course_id
        type: string
      - description: Maximum price
        in: query
        name: max_price
        type: number
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Days Per Week
        in: query
        name: days_per_week
        type: integer
      - description: Lesson Hours
        in: query
        name: lesson_hours
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.GetListCourseItemRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get list of course items
      tags:
      - courseItems
  /courseItems/update:
    put:
      consumes:
      - application/json
      description: Update course item details
      parameters:
      - description: Course Item ID
        in: query
        name: id
        required: true
        type: string
      - description: Update course item
        in: body
        name: course_item
        required: true
        schema:
          $ref: '#/definitions/proto.CreateCourseItem'
      produces:
      - application/json
      responses:
        "200":
          description: Course item updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a course item
      tags:
      - courseItems
  /courses/create:
    post:
      consumes:
      - application/json
      description: Create a course
      parameters:
      - description: Create course
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/proto.CreateCourse'
      produces:
      - application/json
      responses:
        "200":
          description: Course created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a course
      tags:
      - courses
  /courses/delete:
    delete:
      consumes:
      - application/json
      description: Delete a course by ID
      parameters:
      - description: Course ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Course deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a course
      tags:
      - courses
  /courses/get:
    get:
      consumes:
      - application/json
      description: Get a course by its ID
      parameters:
      - description: Course ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.CourseRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a course by ID
      tags:
      - courses
  /courses/list:
    get:
      consumes:
      - application/json
      description: Get a list of courses with optional filtering
      parameters:
      - description: Duration filter
        in: query
        name: duration
        type: number
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.GetListCourseRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all courses
      tags:
      - courses
  /courses/update:
    put:
      consumes:
      - application/json
      description: Update a course by ID
      parameters:
      - description: Course ID
        in: query
        name: id
        required: true
        type: string
      - description: Update course details
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/proto.CreateCourse'
      produces:
      - application/json
      responses:
        "200":
          description: Course updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a course
      tags:
      - courses
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"